Replace all css files to separate folder
