+++ src/assets/components/CategoryFilter.css
@@ -1,0 +1,3 @@
+.categories-list {
+  border: 1px solid grey;
+}
